REFACTOR: Remove bitdubai from wallet redeem point wallet plugin. #5809
