+++ DAP/android/reference_wallet/fermat-dap-android-reference-wallet-redeem-point-bitdubai/src/main/java/com/bitdubai/fermat_dap_android_wallet_redeem_point_bitdubai/adapters/AssetDetailTransactionAdapter.java
@@ -8,7 +8,6 @@ import com.bitdubai.fermat_dap_android_wallet_redeem_point_bitdubai.R;
 import com.bitdubai.fermat_dap_android_wallet_redeem_point_bitdubai.holders.AssetDetailTransactionHolder;
 import com.bitdubai.fermat_dap_android_wallet_redeem_point_bitdubai.models.Transaction;
 import com.bitdubai.fermat_dap_api.layer.dap_module.wallet_asset_redeem_point.interfaces.AssetRedeemPointWalletSubAppModule;
-import com.bitdubai.fermat_dap_plugin.layer.module.wallet.redeem.point.developer.bitdubai.version_1.structure.AssetRedeemPointWalletModuleManager;
 
 import java.util.List;
