+++ DAP/library/core/fermat-dap-core/src/main/java/com/bitdubai/fermat_dap_core/layer/negotiation_transaction/negotiation_direct_sell/NegotiationDirectSellPluginSubsystem.java
@@ -4,7 +4,7 @@ import com.bitdubai.fermat_api.layer.all_definition.common.system.utils.PluginRe
 import com.bitdubai.fermat_api.layer.all_definition.enums.Plugins;
 import com.bitdubai.fermat_core_api.layer.all_definition.system.abstract_classes.AbstractPluginSubsystem;
 import com.bitdubai.fermat_core_api.layer.all_definition.system.exceptions.CantStartSubsystemException;
-import com.bitdubai.fermat_dap_plugin.layer.negotiation_transaction.direct_sell_negotiation.developer.bitdubai.DeveloperBitDubai;
+import org.fermat.fermat_dap_plugin.layer.negotiation_transaction.direct_sell_negotiation.developer.bitdubai.DeveloperBitDubai;
 
 /**
  * Created by Víctor A. Mars M. (marsvicam@gmail.com) on 10/02/16.
