+++ DAP/library/core/fermat-dap-core/src/main/java/com/bitdubai/fermat_dap_core/layer/wallet/asset_user/AssetUserPluginSubsystem.java
@@ -4,7 +4,7 @@ import com.bitdubai.fermat_core_api.layer.all_definition.system.abstract_classes
 import com.bitdubai.fermat_core_api.layer.all_definition.system.exceptions.CantStartSubsystemException;
 import com.bitdubai.fermat_api.layer.all_definition.common.system.utils.PluginReference;
 import com.bitdubai.fermat_api.layer.all_definition.enums.Plugins;
-import com.bitdubai.fermat_dap_plugin.layer.wallet.asset.user.developer.bitdubai.DeveloperBitDubai;
+import org.fermat.fermat_dap_plugin.layer.wallet.asset.user.developer.bitdubai.DeveloperBitDubai;
 
 /**
  * Created by lnacosta - (laion.cj91@gmail.com) on 11/11/2015.
