--- DAP/plugin/wallet/fermat-dap-plugin-wallet-redeem-point-wallet-bitdubai/src/main/java/com/bitdubai/fermat_dap_plugin/layer/wallet/wallet/redeem/point/developer/bitdubai/version_1/structure/functional/AssetRedeemPointWalletBalance.java
@@ -1,77 +1,0 @@
-package com.bitdubai.fermat_dap_plugin.layer.wallet.wallet.redeem.point.developer.bitdubai.version_1.structure.functional;
-
-import com.bitdubai.fermat_dap_api.layer.all_definition.digital_asset.DigitalAsset;
-import com.bitdubai.fermat_dap_api.layer.dap_wallet.asset_redeem_point.interfaces.AssetRedeemPointWalletList;
-
-/**
- * Created by franklin on 15/10/15.
- */
-public class AssetRedeemPointWalletBalance implements AssetRedeemPointWalletList {
-
-    DigitalAsset digitalAsset;
-    long quantityBookBalance;
-    long quantityAvailableBalance;
-    long availableBalance;
-    long bookBalance;
-
-    public AssetRedeemPointWalletBalance(DigitalAsset digitalAsset, long quantityBookBalance, long quantityAvailableBalance, long availableBalance, long bookBalance) {
-        this.digitalAsset = digitalAsset;
-        this.quantityBookBalance = quantityBookBalance;
-        this.quantityAvailableBalance = quantityAvailableBalance;
-        this.availableBalance = availableBalance;
-        this.bookBalance = bookBalance;
-    }
-
-    public AssetRedeemPointWalletBalance() {
-    }
-
-    @Override
-    public long getQuantityAvailableBalance() {
-        return quantityAvailableBalance;
-    }
-
-    @Override
-    public void setQuantityAvailableBalance(long quantityAvailableBalance) {
-        this.quantityAvailableBalance = quantityAvailableBalance;
-    }
-
-    @Override
-    public DigitalAsset getDigitalAsset() {
-        return digitalAsset;
-    }
-
-    @Override
-    public void setDigitalAsset(DigitalAsset digitalAsset) {
-        this.digitalAsset = digitalAsset;
-    }
-
-    @Override
-    public long getAvailableBalance() {
-        return availableBalance;
-    }
-
-    @Override
-    public void setAvailableBalance(long availableBalance) {
-        this.availableBalance = availableBalance;
-    }
-
-    @Override
-    public long getBookBalance() {
-        return bookBalance;
-    }
-
-    @Override
-    public void setBookBalance(long bookBalance) {
-        this.bookBalance = bookBalance;
-    }
-
-    @Override
-    public long getQuantityBookBalance() {
-        return quantityBookBalance;
-    }
-
-    @Override
-    public void setQuantityBookBalance(long quantityBookBalance) {
-        this.quantityBookBalance = quantityBookBalance;
-    }
-}
