--- DAP/plugin/wallet/fermat-dap-plugin-wallet-redeem-point-wallet-bitdubai/src/main/java/com/bitdubai/fermat_dap_plugin/layer/wallet/wallet/redeem/point/developer/bitdubai/version_1/structure/functional/AssetRedeemPointWalletBalanceImpl.java
@@ -1,58 +1,0 @@
-package com.bitdubai.fermat_dap_plugin.layer.wallet.wallet.redeem.point.developer.bitdubai.version_1.structure.functional;
-
-import com.bitdubai.fermat_api.layer.all_definition.enums.WalletsPublicKeys;
-import com.bitdubai.fermat_api.layer.osa_android.broadcaster.Broadcaster;
-import com.bitdubai.fermat_api.layer.osa_android.broadcaster.BroadcasterType;
-import com.bitdubai.fermat_dap_api.layer.all_definition.DAPConstants;
-import com.bitdubai.fermat_dap_api.layer.dap_wallet.asset_issuer_wallet.exceptions.CantCalculateBalanceException;
-import com.bitdubai.fermat_dap_api.layer.dap_wallet.asset_issuer_wallet.exceptions.CantRegisterCreditException;
-import com.bitdubai.fermat_dap_api.layer.dap_wallet.asset_issuer_wallet.exceptions.CantRegisterDebitException;
-import com.bitdubai.fermat_dap_api.layer.dap_wallet.asset_redeem_point.interfaces.AssetRedeemPointWalletList;
-import com.bitdubai.fermat_dap_api.layer.dap_wallet.asset_redeem_point.interfaces.AssetRedeemPointWalletTransactionRecord;
-import com.bitdubai.fermat_dap_api.layer.dap_wallet.common.enums.BalanceType;
-import com.bitdubai.fermat_dap_plugin.layer.wallet.wallet.redeem.point.developer.bitdubai.version_1.structure.database.AssetRedeemPointWalletDao;
-
-import java.util.List;
-
-/**
- * Created by franklin on 16/10/15.
- */
-public class AssetRedeemPointWalletBalanceImpl implements com.bitdubai.fermat_dap_api.layer.dap_wallet.asset_redeem_point.interfaces.AssetRedeemPointWalletBalance {
-
-    private AssetRedeemPointWalletDao assetRedeemPointWalletDao;
-    private Broadcaster broadcaster;
-
-    /**
-     * Constructor.
-     */
-    public AssetRedeemPointWalletBalanceImpl(AssetRedeemPointWalletDao assetRedeemPointWalletDao, Broadcaster broadcaster) {
-        this.assetRedeemPointWalletDao = assetRedeemPointWalletDao;
-        this.broadcaster = broadcaster;
-    }
-
-    @Override
-    public long getBalance() throws CantCalculateBalanceException {
-        return assetRedeemPointWalletDao.getAvailableBalance();
-    }
-
-    @Override
-    public List<AssetRedeemPointWalletList> getAssetIssuerWalletBalances() throws CantCalculateBalanceException {
-        return assetRedeemPointWalletDao.getBalanceByAssets();
-    }
-
-    @Override
-    public void debit(AssetRedeemPointWalletTransactionRecord assetRedeemPointWalletTransactionRecord, BalanceType balanceType) throws CantRegisterDebitException {
-        assetRedeemPointWalletDao.addDebit(assetRedeemPointWalletTransactionRecord, balanceType);
-
-        broadcaster.publish(BroadcasterType.UPDATE_VIEW, DAPConstants.DAP_UPDATE_VIEW_ANDROID);
-        broadcaster.publish(BroadcasterType.NOTIFICATION_SERVICE, WalletsPublicKeys.DAP_REDEEM_WALLET.getCode(), "ASSET-REDEEM-DEBIT_" + "Name: " + assetRedeemPointWalletTransactionRecord.getDigitalAsset().getName() + " - Balance: " + balanceType.getCode());
-    }
-
-    @Override
-    public void credit(AssetRedeemPointWalletTransactionRecord assetRedeemPointWalletTransactionRecord, BalanceType balanceType) throws CantRegisterCreditException {
-        assetRedeemPointWalletDao.addCredit(assetRedeemPointWalletTransactionRecord, balanceType);
-
-        broadcaster.publish(BroadcasterType.UPDATE_VIEW, DAPConstants.DAP_UPDATE_VIEW_ANDROID);
-        broadcaster.publish(BroadcasterType.NOTIFICATION_SERVICE, WalletsPublicKeys.DAP_REDEEM_WALLET.getCode(), "ASSET-REDEEM-CREDIT_" + "Name: " + assetRedeemPointWalletTransactionRecord.getDigitalAsset().getName() + " - Balance: " + balanceType.getCode());
-    }
-}
