+++ DAP/plugin/wallet/fermat-dap-plugin-wallet-redeem-point-wallet/build.gradle
